# Conflicts:
#	moorestech_server/Assets/Scripts/Server.Protocol/PacketResponse/PlaceBlockFromHotBarProtocol.cs

+++ moorestech_server/Assets/Scripts/Game.Block/Blocks/Util/MachineCurrentPowerToSubMillSecond.cs
@@ -8,13 +8,13 @@ namespace Game.Block.Blocks.Util
     /// </summary>
     public static class MachineCurrentPowerToSubMillSecond
     {
-        public static int GetSubMillSecond(ElectricPower currentPower, ElectricPower requiredPower)
+        public static double GetSubMillSecond(ElectricPower currentPower, ElectricPower requiredPower)
         {
             //必要電力が0の時はそのフレームの時間を返す
-            if (requiredPower.AsPrimitive() == 0) return (int)GameUpdater.UpdateMillSecondTime;
+            if (requiredPower.AsPrimitive() == 0) return GameUpdater.UpdateMillSecondTime;
             //現在の電力量を必要電力で割った割合で、そのフレームの時間を返す
             //例えば、必要電力が100、現在の電力が50だったら、そのフレームの半分の時間を返すことで、機械の速度を半分にする
-            return (int)(GameUpdater.UpdateMillSecondTime * (currentPower.AsPrimitive() / (double)requiredPower.AsPrimitive()));
+            return GameUpdater.UpdateMillSecondTime * (currentPower.AsPrimitive() / (double)requiredPower.AsPrimitive());
         }
     }
 }
